@@ -73,45 +78,44 @@
     "react-hot-loader": "^1.3.0",
     "react-router": "^2.6.1",
     "react-router-scroll": "^0.3.1",
-    "sass-loader": "^3.1.2",
+    "sass-loader": "^4.0.0",
     "slash": "^1.0.0",
     "static-site-generator-webpack-plugin": "^2.1.0",
-    "style-loader": "^0.13.0",
-    "toml": "^2.2.2",
+    "style-loader": "^0.13.2",
+    "toml": "^2.3.2",
     "toml-loader": "^1.0.0",
     "tracer": "^0.8.3",
     "ts-loader": "^1.2.2",
-    "url-loader": "^0.5.7",
+    "url-loader": "^0.5.8",
     "webpack": "^1.13.2",
     "webpack-configurator": "^0.3.0",
-    "webpack-hot-middleware": "^2.12.2",
+    "webpack-hot-middleware": "^2.17.1",
     "webpack-require": "0.0.16",
     "yaml-loader": "^0.4.0"
   },
   "devDependencies": {
-    "ava": "^0.16.0",
+    "ava": "0.16",
     "ava-http": "^0.2.1",
-    "babel-cli": "^6.11.4",
-    "babel-eslint": "^6.1.2",
+    "babel-cli": "^6.24.0",
+    "babel-eslint": "^7.2.0",
     "babel-plugin-lodash": "^3.2.8",
     "babel-plugin-transform-flow-strip-types": "^6.8.0",
-    "babel-register": "^6.11.6",
-    "bluebird": "^3.4.1",
-    "chalk": "^1.1.3",
-    "cheerio": "^0.20.0",
-    "eslint": "^2.11.1",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-ava": "^2.5.0",
+    "babel-register": "^6.24.0",
+    "bluebird": "^3.5.0",
+    "cheerio": "^0.22.0",
+    "eslint": "^3.18.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-plugin-ava": "^4.2.0",
     "eslint-plugin-flow-vars": "^0.5.0",
-    "eslint-plugin-flowtype": "^2.7.1",
-    "eslint-plugin-import": "^1.13.0",
-    "eslint-plugin-jsx-a11y": "1.5.3",
-    "eslint-plugin-react": "^5.2.2",
-    "flow-bin": "^0.30.0",
+    "eslint-plugin-flowtype": "^2.30.4",
+    "eslint-plugin-import": "2.2.x",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.10.3",
+    "flow-bin": "^0.42.0",
     "iflow-debug": "^1.0.15",
     "iflow-lodash": "^1.1.23",
     "iflow-react-router": "^1.1.17",
-    "nyc": "^7.0.0"
+    "nyc": "^10.1.2"
   },
   "engines": {
     "node": ">0.12.0"
