@@ -1,8 +1,8 @@
 /* @flow weak */
-const toml = require('toml')
-const loaderUtils = require('loader-utils')
-const path = require('path')
-const globPages = require('../../utils/glob-pages')
+import toml from 'toml'
+import loaderUtils from 'loader-utils'
+import path from 'path'
+import globPages from '../../utils/glob-pages'
 
 module.exports = function (source) {
   this.cacheable()
@@ -15,6 +15,6 @@ module.exports = function (source) {
   value.relativePath = path.relative('.', directory)
   globPages(directory, (err, pagesData) => {
     value.pages = pagesData
-    return callback(null, `module.exports = ${JSON.stringify(value, void 0, '\t')}`)
+    return callback(null, `module.exports = ${JSON.stringify(value, null, '\t')}`)
   })
 }
