@@ -2,7 +2,7 @@ const program = require('commander')
 const packageJson = require('../../package.json')
 const fs = require('fs-extra')
 const path = require('path')
-import _ from 'lodash'
+const _ = require('lodash')
 
 // Copy our load-context function to root of site in a dot file.
 const gatsbyFile = `${__dirname}/../utils/load-context.js`
@@ -24,7 +24,7 @@ program.command('develop')
   .description('Start development server. Watches files and rebuilds and hot reloads if something changes') // eslint-disable-line max-len
   .option('-H, --host <url>',
           `Set host. Defaults to ${defaultHost}`,
-          defaultHost
+          defaultHost,
          )
   .option('-p, --port <port>', 'Set port. Defaults to 8000', '8000')
   .option('-o, --open', 'Open the site in your browser for you.')
@@ -62,7 +62,7 @@ program.command('serve-build')
   .description('Serve built site.')
   .option('-H, --host <url>',
           `Set host. Defaults to ${defaultHost}`,
-          defaultHost
+          defaultHost,
          )
   .option('-p, --port <port>', 'Set port. Defaults to 8000', '8000')
   .option('-o, --open', 'Open the site in your browser for you.')
