@@ -1,5 +1,7 @@
 /* @flow weak */
-import _ from 'lodash'
+import filter from 'lodash/filter'
+import last from 'lodash/last'
+import sortBy from 'lodash/sortBy'
 import { prefixLink } from './gatsby-helpers'
 
 function requireComponent (moduleReq) {
@@ -19,8 +21,9 @@ function requireComponent (moduleReq) {
 module.exports = (files, pagesReq) => {
   // Remove files that start with an underscore as this indicates
   // the file shouldn't be turned into a page.
-  const pages = _.filter(files, (file) => file.file.name.slice(0, 1) !== '_')
-  const templates = _.filter(files, (file) =>
+  const pages = filter(files, file => file.file.name.slice(0, 1) !== '_')
+  const templates = filter(files, file =>
+    // eslint-disable-next-line comma-dangle
     file.file.name === '_template'
   )
 
@@ -46,28 +49,30 @@ module.exports = (files, pagesReq) => {
   // 3. For each index file paired with a template, create a default route
   // 4. Create normal routes for each remaining file under the appropriate
   // template
-  const templatesWithoutRoot = _.filter(files, (file) =>
+  const templatesWithoutRoot = filter(files, file =>
+    // eslint-disable-next-line comma-dangle
     file.file.name === '_template' && file.file.dirname !== ''
   )
 
   // Find the parent template for each template file and create a route
   // with it.
   templatesWithoutRoot.forEach((templateFile) => {
-    let parentTemplates = _.filter(templatesWithoutRoot, (template) =>
+    let parentTemplates = filter(templatesWithoutRoot, template =>
+      // eslint-disable-next-line comma-dangle
       templateFile.requirePath.indexOf(template.file.dirname) === 0
     )
     // Sort parent templates by directory length. In cases
     // where a template has multiple parents
     // e.g. /_template.js/blog/_template.js/archive/_template.js
     // we want to nest this template under its most immediate parent.
-    parentTemplates = _.sortBy(parentTemplates, (template) => {
+    parentTemplates = sortBy(parentTemplates, (template) => {
       if (template) {
         return template.file.dirname.length
       } else {
         return 0
       }
     })
-    const parentTemplateFile = _.last(parentTemplates)
+    const parentTemplateFile = last(parentTemplates)
     let parentRoute
     if (parentTemplateFile) {
       parentRoute = templatesHash[parentTemplateFile.file.dirname]
@@ -97,7 +102,13 @@ module.exports = (files, pagesReq) => {
   const staticFileTypes = [
     'ipynb',
     'md',
+    'rmd',
+    'mkd',
+    'mkdn',
+    'mdwn',
+    'mdown',
     'markdown',
+    'litcoffee',
     'html',
     'json',
     'yaml',
@@ -132,24 +143,19 @@ module.exports = (files, pagesReq) => {
       handler = wrappers[page.file.ext]
       page.data = pagesReq(`./${page.requirePath}`)
     } else if (reactComponentFileTypes.indexOf(page.file.ext) !== -1) {
-      handler = requireComponent(pagesReq(`./${page.requirePath}`))
-      page.data = (() => {
-        if (pagesReq(`./${page.requirePath}`).metadata) {
-          return pagesReq(`./${page.requirePath}`).metadata()
-        } else {
-          return {}
-        }
-      })()
+      handler = pagesReq(`./${page.requirePath}`)
+      page.data = (page.data === undefined) ? {} : page.data
     }
 
     // Determine parent template for page.
-    const parentTemplates = _.filter(templatesWithoutRoot, (templateFile) =>
+    const parentTemplates = filter(templatesWithoutRoot, templateFile =>
+      // eslint-disable-next-line comma-dangle
       page.requirePath.indexOf(templateFile.file.dirname) === 0
     )
 
-    const sortedParentTemplates = _.sortBy(parentTemplates, (route) => route.file.dirname.length)
+    const sortedParentTemplates = sortBy(parentTemplates, route => route.file.dirname.length)
 
-    const parentTemplateFile = _.last(sortedParentTemplates)
+    const parentTemplateFile = last(sortedParentTemplates)
     let parentRoute
     if (parentTemplateFile) {
       parentRoute = templatesHash[parentTemplateFile.file.dirname]
