@@ -1,11 +1,11 @@
 /* @flow weak */
 import { config } from 'config'
 import invariant from 'invariant'
-import _ from 'lodash'
+import isString from 'lodash/isString'
 
 function isDataURL (s) {
   // Regex from https://gist.github.com/bgrins/6194623#gistcomment-1671744
-  // eslint-disable-next-line max-len
+  // eslint-disable-next-line
   const regex = /^\s*data:([a-z]+\/[a-z0-9\-\+]+(;[a-z\-]+=[a-z0-9\-]+)?)?(;base64)?,[a-z0-9!\$&',\(\)\*\+,;=\-\._~:@\/\?%\s]*\s*$/i
   return !!s.match(regex)
 }
@@ -20,7 +20,7 @@ const prefixLink = (_link) => {
     You're trying to build your site with links prefixed
     but you haven't set 'linkPrefix' in your config.toml.
     `
-    invariant(_.isString(config.linkPrefix), invariantMessage)
+    invariant(isString(config.linkPrefix), invariantMessage)
 
     // When url-loader is used with the limit option, some of images are encoded to
     // base64 data url, this check prevent such links from being prefixed.
