@@ -14,7 +14,7 @@ console.info(defaultMessage)
 
 function HTML (props) {
   return (
-    <html>
+    <html lang="en">
       <head>
         <meta charSet="utf-8" />
         <meta httpEquiv="X-UA-Compatible" content="IE=edge" />
@@ -31,6 +31,7 @@ function HTML (props) {
   )
 }
 
-HTML.propTypes = { body: PropTypes.any }
+HTML.propTypes = { body: PropTypes.node }
+HTML.defaultProps = { body: '' }
 
 module.exports = HTML
