@@ -14,6 +14,7 @@ function template (props) {
   return <div>{props.children}</div>
 }
 
-template.propTypes = { children: PropTypes.any }
+template.propTypes = { children: PropTypes.node }
+template.defaultProps = { children: null }
 
 module.exports = template
