@@ -5,11 +5,19 @@ module.exports = React.createClass({
   propTypes: {
     route: PropTypes.shape({
       page: PropTypes.shape({
-        data: PropTypes.shape({
-          body: PropTypes.string.isRequired,
-        }),
+        data: PropTypes.object,
       }),
     }),
+  },
+
+  getDefaultProps () {
+    return {
+      route: {
+        page: {
+          data: {},
+        },
+      },
+    }
   },
 
   render () {
