@@ -3,6 +3,8 @@ import frontMatter from 'front-matter'
 import markdownIt from 'markdown-it'
 import hljs from 'highlight.js'
 import objectAssign from 'object-assign'
+import path from 'path'
+import loaderUtils from 'loader-utils'
 
 const highlight = (str, lang) => {
   if ((lang !== null) && hljs.getLanguage(lang)) {
@@ -20,17 +22,29 @@ const highlight = (str, lang) => {
   return ''
 }
 
-const md = markdownIt({
+const md = (linkPrefix, shouldPrefix) => markdownIt({
   html: true,
   linkify: true,
   typographer: true,
   highlight,
+  replaceLink: (link) => {
+    if (shouldPrefix && path.isAbsolute(link)) {
+      return linkPrefix + link
+    }
+    return link
+  },
 })
+  .use(require('markdown-it-replace-link'))
 
 module.exports = function (content) {
   this.cacheable()
+
+  const query = loaderUtils.parseQuery(this.query)
+  const linkPrefix = query.config.linkPrefix || ''
+  const shouldPrefix = query.shouldPrefix
+
   const meta = frontMatter(content)
-  const body = md.render(meta.body)
+  const body = md(linkPrefix, shouldPrefix).render(meta.body)
   const result = objectAssign({}, meta.attributes, {
     body,
   })
