@@ -26,6 +26,7 @@ function defaultConfig () {
  */
 function resolvePlugin (pluginName, directory, type) {
   const gatsbyPath = path.resolve(__dirname, '..', '..')
+
   const plugin = resolve(`babel-${type}-${pluginName}`, directory) ||
     resolve(`babel-${type}-${pluginName}`, gatsbyPath) ||
     resolve(pluginName, directory) ||
@@ -54,18 +55,23 @@ function resolvePlugin (pluginName, directory, type) {
  * regardless of where they are located.
  */
 function normalizeConfig (config, directory) {
-  const normalizedConfig = {
-    presets: [],
-    plugins: [],
-  }
+  const normalizedConfig = { presets: [], plugins: [] }
 
   const presets = config.presets || []
-  presets.forEach(preset => {
-    normalizedConfig.presets.push(resolvePlugin(preset, directory, 'preset'))
+  presets.forEach((preset) => {
+    let normalizedPreset
+
+    if (_.isArray(preset)) {
+      normalizedPreset = [resolvePlugin(preset[0], directory, 'preset'), preset[1]]
+    } else {
+      normalizedPreset = resolvePlugin(preset, directory, 'preset')
+    }
+
+    normalizedConfig.presets.push(normalizedPreset)
   })
 
   const plugins = config.plugins || []
-  plugins.forEach(plugin => {
+  plugins.forEach((plugin) => {
     let normalizedPlugin
 
     if (_.isArray(plugin)) {
@@ -123,15 +129,13 @@ function findBabelPackage (directory) {
 export default function babelConfig (program, stage) {
   const { directory } = program
 
-  const babelrc = findBabelrc(directory) ||
-    findBabelPackage(directory) ||
-    defaultConfig()
+  const babelrc = findBabelrc(directory) || findBabelPackage(directory) || defaultConfig()
 
   if (stage === 'develop') {
     babelrc.presets.unshift('react-hmre')
   }
 
-  if (!babelrc.hasOwnProperty('cacheDirectory')) {
+  if (!_.has(babelrc, 'cacheDirectory')) {
     babelrc.cacheDirectory = true
   }
 
