@@ -20,7 +20,7 @@ export default function pathResolver (pageData, parsedPath) {
    * Then enforces a starting and trailing slash where applicable on the url.
    */
   function hardcodedPath () {
-    if (_.isUndefined(pageData.path)) return void 0
+    if (_.isUndefined(pageData.path)) return undefined
 
     let pagePath = pageData.path
 
