@@ -1,6 +1,6 @@
 /* @flow weak */
+/* eslint import/imports-first: 0 */
 require('node-cjsx').transform()
-import detect from 'detect-port'
 import Hapi from 'hapi'
 import Boom from 'boom'
 import React from 'react'
@@ -14,14 +14,8 @@ import WebpackPlugin from 'hapi-webpack-plugin'
 import opn from 'opn'
 import fs from 'fs'
 import glob from 'glob'
-import rl from 'readline'
 import chalk from 'chalk'
-import getProcessForPort from './getProcessForPort'
-
-const rlInterface = rl.createInterface({
-  input: process.stdin,
-  output: process.stdout,
-})
+import startWithDynamicPort from './startWithDynamicPort'
 
 import globPages from './glob-pages'
 import webpackConfig from './webpack.config'
@@ -82,7 +76,7 @@ function startServer (program) {
             const htmlElement = React.createElement(
               HTML, {
                 body: devNoScript,
-              }
+              },
             )
             let html = ReactDOMServer.renderToStaticMarkup(htmlElement)
             html = `<!DOCTYPE html>\n${html}`
@@ -111,7 +105,7 @@ function startServer (program) {
 
         // Try to map the url path to match an actual path of a file on disk.
         const parsed = parsePath(request.path)
-        const page = _.find(pages, (p) => p.path === (`${parsed.dirname}/`))
+        const page = _.find(pages, p => p.path === (`${parsed.dirname}/`))
 
         let absolutePath = `${program.directory}/pages`
         let path
@@ -183,7 +177,7 @@ function startServer (program) {
           if (e) {
             if (e.code === 'EADDRINUSE') {
               // eslint-disable-next-line max-len
-              console.log(chalk.red(`Unable to start Gatsby on port ${program.port} as there's already a process listing on that port.`))
+              console.log(chalk.red(`Unable to start Gatsby on port ${program.port} as there's already a process listening on that port.`))
             } else {
               console.log(chalk.red(e))
             }
@@ -205,32 +199,4 @@ function startServer (program) {
   })
 }
 
-module.exports = (program) => {
-  const port = typeof program.port === 'string'
-    ? parseInt(program.port, 10)
-    : program.port
-
-  const existingProcess = getProcessForPort(port)
-
-  detect(port, (err, _port) => {
-    if (err) {
-      console.error(err)
-      process.exit()
-    }
-
-    if (port !== _port) {
-      // eslint-disable-next-line max-len
-      const question = chalk.yellow(`Something is already running on port ${port}.\n${(existingProcess) ? ` Probably:\n  ${existingProcess}\n` : ''}\nWould you like to run the app at another port instead? [Y/n]`)
-
-      return rlInterface.question(question, (answer) => {
-        if (answer.length === 0 || answer.match(/^yes|y$/i)) {
-          program.port = _port // eslint-disable-line no-param-reassign
-        }
-
-        return startServer(program)
-      })
-    }
-
-    return startServer(program)
-  })
-}
+module.exports = startWithDynamicPort(startServer)
