@@ -34,14 +34,14 @@ function getProcessCommand (processId, processDirectory) {
 
   if (isProcessAReactApp(command)) {
     const packageName = getPackageNameInDirectory(processDirectory)
-    return (packageName) ? '${packageName}\n' : command
+    return (packageName) ? `${packageName}\n` : command
   } else {
     return command
   }
 }
 
 function getDirectoryOfProcessById (processId) {
-  return execSync(`lsof -p ${processId} | grep cwd | awk \'{print $9}\'`, execOptions).trim()
+  return execSync(`lsof -p ${processId} | grep cwd | awk '{print $9}'`, execOptions).trim()
 }
 
 function getProcessForPort (port) {
