@@ -1,12 +1,12 @@
 /* @flow weak */
 import glob from 'glob'
 import buildPage from './build-page'
-const debug = require('debug')('gatsby:glob')
-
 import pageFileTypes from './page-file-types'
 
+const debug = require('debug')('gatsby:glob')
+
 function globQuery (directory) {
-  const fileGlobQuery = pageFileTypes.map((type) => `*.${type}`)
+  const fileGlobQuery = pageFileTypes.map(type => `*.${type}`)
   const joinedFileQuery = fileGlobQuery.join('|')
   return `${directory}/pages/**/?(${joinedFileQuery})`
 }
