@@ -30,7 +30,7 @@ const install = (rootPath, callback) => {
   })
 }
 
-const ignored = (path) => !/^\.(git|hg)$/.test(sysPath.basename(path))
+const ignored = path => !/^\.(git|hg)$/.test(sysPath.basename(path))
 
 // Copy starter from file system.
 //
@@ -78,7 +78,7 @@ const clone = (address, rootPath, callback) => {
   const url = gitHubRe.test(address) ?
     `git://github.com/${address.replace(gitHubRe, '')}.git` : address
   logger.log(`Cloning git repo ${url} to ${rootPath}...`)
-  const cmd = `git clone ${url} ${rootPath}`
+  const cmd = `git clone --depth=1 ${url} ${rootPath}`
   exec(cmd, (error, stdout, stderr) => {
     if (error !== null) {
       return callback(new Error(`Git clone error: ${stderr.toString()}`))
