@@ -5,8 +5,14 @@ const pageFileTypes = [
   'tsx',
   'jsx',
   'js',
-  'markdown',
   'md',
+  'rmd',
+  'mkd',
+  'mkdn',
+  'mdwn',
+  'mdown',
+  'markdown',
+  'litcoffee',
   'html',
   'json',
   'yaml',
