@@ -5,9 +5,10 @@ import fs from 'fs-extra'
 import async from 'async'
 import parsePath from 'parse-filepath'
 import _ from 'lodash'
+import globPages from './glob-pages'
+
 const debug = require('debug')('gatsby:post-build')
 
-import globPages from './glob-pages'
 
 module.exports = (program, cb) => {
   const directory = program.directory
@@ -52,7 +53,7 @@ module.exports = (program, cb) => {
       }
 
       newPath = `${directory}/public/${newPath}`
-      return fs.copy(file, newPath, (error) => callback(error))
+      return fs.copy(file, newPath, error => callback(error))
     }
 
     // Copy static assets to public folder.
@@ -60,7 +61,9 @@ module.exports = (program, cb) => {
     const globString = `${directory}/pages/**/?(${assetTypes})`
     return glob(globString, { follow: true }, (e, files) =>
       async.map(files, copy, (error, results) =>
+        // eslint-disable-next-line comma-dangle
         cb(error, results)
+      // eslint-disable-next-line comma-dangle
       )
     )
   })
