@@ -1,15 +1,8 @@
 /* @flow weak */
-import detect from 'detect-port'
 import Hapi from 'hapi'
 import opn from 'opn'
-import rl from 'readline'
 import chalk from 'chalk'
-import getProcessForPort from './getProcessForPort'
-
-const rlInterface = rl.createInterface({
-  input: process.stdin,
-  output: process.stdout,
-})
+import startWithDynamicPort from './startWithDynamicPort'
 
 const debug = require('debug')('gatsby:application')
 
@@ -42,7 +35,7 @@ function startServer (program, launchPort) {
     if (e) {
       if (e.code === 'EADDRINUSE') {
         // eslint-disable-next-line max-len
-        console.log(chalk.red(`Unable to start Gatsby on port ${serverPort} as there's already a process listing on that port.`))
+        console.log(chalk.red(`Unable to start Gatsby on port ${serverPort} as there's already a process listening on that port.`))
       } else {
         console.log(chalk.red(e))
       }
@@ -61,29 +54,4 @@ function startServer (program, launchPort) {
   })
 }
 
-module.exports = (program) => {
-  const existingProcess = getProcessForPort(program.port)
-
-  detect(program.port, (err, _port) => {
-    if (err) {
-      console.error(err)
-      process.exit()
-    }
-
-    if (program.port !== _port) {
-      // eslint-disable-next-line max-len
-      const question = chalk.yellow(`Something is already running on port ${program.port}.\n${(existingProcess) ? ` Probably:\n  ${existingProcess}\n` : ''}\nWould you like to run the app at another port instead? [Y/n]`)
-
-      return rlInterface.question(question, (answer) => {
-        let launchPort = program.port
-        if (answer.length === 0 || answer.match(/^yes|y$/i)) {
-          launchPort = _port
-        }
-
-        return startServer(program, launchPort)
-      })
-    }
-
-    return startServer(program)
-  })
-}
+module.exports = startWithDynamicPort(startServer)
