@@ -4,9 +4,11 @@ import { renderToString, renderToStaticMarkup } from 'react-dom/server'
 import { match, RouterContext } from 'react-router'
 import createRoutes from 'create-routes'
 import Html from 'html'
-const loadContext = require('.gatsby-context')
 import { pages, config } from 'config'
 import { prefixLink } from '../isomorphic/gatsby-helpers'
+import { wrapRootComponent } from 'gatsby-ssr'
+
+const loadContext = require('.gatsby-context')
 
 let routes
 loadContext((pagesReq) => {
@@ -19,17 +21,25 @@ module.exports = (locals, callback) => {
       console.log(error)
       callback(error)
     } else if (renderProps) {
-      const component = <RouterContext {...renderProps} />
+      const componentToWrap = () => (<RouterContext {...renderProps} />)
       let body
+      let component
+
+      if (wrapRootComponent) {
+        component = wrapRootComponent(componentToWrap)
+      } else {
+        component = componentToWrap
+      }
 
       // If we're not generating a SPA for production, eliminate React IDs.
       if (config.noProductionJavascript) {
-        body = renderToStaticMarkup(component)
+        body = renderToStaticMarkup(component())
       } else {
-        body = renderToString(component)
+        body = renderToString(component())
       }
 
       const html = `<!DOCTYPE html>\n ${renderToStaticMarkup(
+        // eslint-disable-next-line comma-dangle
         <Html body={body} {...renderProps} />
       )}`
       callback(null, html)
