@@ -71,7 +71,7 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
           publicPath,
         }
       case 'build-html':
-        // A temp file required by static-site-generator-plugin. See plugins() below.
+        // A temp file required by static-site-generator-plugin. See plugins function.
         // Deleted by build-html.js, since it's not needed for production.
         return {
           path: `${directory}/public`,
@@ -114,57 +114,37 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
     }
   }
 
-  function plugins () {
+  function plugins (config) {
+    config.plugin('define', webpack.DefinePlugin, [{
+      'process.env': {
+        NODE_ENV: JSON.stringify(process.env.NODE_ENV ? process.env.NODE_ENV : 'development'),
+      },
+      __PREFIX_LINKS__: program.prefixLinks,
+    }])
+
     switch (stage) {
       case 'develop':
-        return [
-          new webpack.optimize.OccurenceOrderPlugin(),
-          new webpack.HotModuleReplacementPlugin(),
-          new webpack.NoErrorsPlugin(),
-          new webpack.DefinePlugin({
-            'process.env': {
-              NODE_ENV: JSON.stringify(process.env.NODE_ENV ? process.env.NODE_ENV : 'development'),
-            },
-            __PREFIX_LINKS__: program.prefixLinks,
-          }),
-        ]
+        config.plugin('optimize-order', webpack.optimize.OccurenceOrderPlugin, null)
+        config.plugin('hot-module-replacement', webpack.HotModuleReplacementPlugin, null)
+        config.plugin('no-errors', webpack.NoErrorsPlugin, null)
+
+        return config
       case 'build-css':
-        return [
-          new webpack.DefinePlugin({
-            'process.env': {
-              NODE_ENV: JSON.stringify(process.env.NODE_ENV ? process.env.NODE_ENV : 'production'),
-            },
-            __PREFIX_LINKS__: program.prefixLinks,
-          }),
-          new ExtractTextPlugin('styles.css'),
-        ]
+        config.plugin('extract-text', ExtractTextPlugin, ['styles.css'])
+
+        return config
       case 'build-html':
-        return [
-          new StaticSiteGeneratorPlugin('render-page.js', routes),
-          new webpack.DefinePlugin({
-            'process.env': {
-              NODE_ENV: JSON.stringify(process.env.NODE_ENV ? process.env.NODE_ENV : 'production'),
-            },
-            __PREFIX_LINKS__: program.prefixLinks,
-          }),
-          new ExtractTextPlugin('styles.css'),
-        ]
+        config.plugin('static-site-generator', StaticSiteGeneratorPlugin, ['render-page.js', routes])
+        config.plugin('extract-text', ExtractTextPlugin, ['styles.css'])
+
+        return config
       case 'build-javascript':
-        return [
-          // Moment.js includes 100s of KBs of extra localization data
-          // by default in Webpack that most sites don't want.
-          // This line disables that.
-          new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
-          new webpack.DefinePlugin({
-            'process.env': {
-              NODE_ENV: JSON.stringify(process.env.NODE_ENV ? process.env.NODE_ENV : 'production'),
-            },
-            __PREFIX_LINKS__: program.prefixLinks,
-          }),
-          new ExtractTextPlugin('styles.css'),
-          new webpack.optimize.DedupePlugin(),
-          new webpack.optimize.UglifyJsPlugin(),
-        ]
+        config.plugin('ignore', webpack.IgnorePlugin, [/^\.\/locale$/, /moment$/])
+        config.plugin('extract-text', ExtractTextPlugin, ['styles.css'])
+        config.plugin('dedupe', webpack.optimize.DedupePlugin, null)
+        config.plugin('uglify', webpack.optimize.UglifyJsPlugin, null)
+
+        return config
       default:
         throw new Error(`The state requested ${stage} doesn't exist.`)
     }
@@ -221,7 +201,7 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
     }
   }
 
-  function module (config) {
+  function loaders (config) {
     // Common config for every env.
     config.loader('cjsx', {
       test: /\.cjsx$/,
@@ -245,8 +225,12 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
       ],
     })
     config.loader('md', {
-      test: /\.md$/,
+      test: /\.(md|rmd|mkdn?|mdwn|mdown|markdown|litcoffee)$/,
       loader: 'markdown',
+      query: {
+        config: siteConfig,
+        shouldPrefix: program.prefixLinks,
+      },
     })
     config.loader('html', {
       test: /\.html$/,
@@ -325,12 +309,12 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
         config.loader('less', {
           test: /\.less/,
           exclude: /\.module\.less$/,
-          loaders: ['style', 'css', 'less'],
+          loaders: ['style', 'css', 'postcss', 'less'],
         })
         config.loader('sass', {
           test: /\.(sass|scss)/,
           exclude: /\.module\.(sass|scss)$/,
-          loaders: ['style', 'css', 'sass'],
+          loaders: ['style', 'css', 'postcss', 'sass'],
         })
 
         // CSS modules
@@ -368,12 +352,12 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
         config.loader('less', {
           test: /\.less/,
           exclude: /\.module\.less$/,
-          loader: ExtractTextPlugin.extract(['css?minimize', 'less']),
+          loader: ExtractTextPlugin.extract(['css?minimize', 'postcss', 'less']),
         })
         config.loader('sass', {
           test: /\.(sass|scss)/,
           exclude: /\.module\.(sass|scss)$/,
-          loader: ExtractTextPlugin.extract(['css?minimize', 'sass']),
+          loader: ExtractTextPlugin.extract(['css?minimize', 'postcss', 'sass']),
         })
 
         // CSS modules
@@ -476,6 +460,13 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
     }
   }
 
+  function module (config) {
+    plugins(config)
+    loaders(config)
+
+    return config
+  }
+
   const config = new Config()
 
   config.merge({
@@ -501,7 +492,6 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
         path.resolve(directory, 'node_modules/gatsby/node_modules'),
       ],
     },
-    plugins: plugins(),
     resolve: resolve(),
   })
 
