@@ -1,12 +1,14 @@
 import test from 'ava'
 import fse from 'fs-extra'
 import Promise from 'bluebird'
+import { gatsby } from '../support'
+
 const fs = Promise.promisifyAll(fse)
 const tmpdir = require('os').tmpdir()
 
-import { gatsby } from '../support'
 
-test('calling gatsby new succesfully creates new site from default starter', async t => {
+// eslint-disable-next-line
+test('calling gatsby new succesfully creates new site from default starter', async (t) => {
   const sitePath = `${tmpdir}/gatsby-default-starter`
   await fs.remove(sitePath)
   const noArgs = await gatsby(['new', sitePath])
