@@ -8,7 +8,7 @@ function programStub (fixture) {
   return { directory }
 }
 
-test('it returns a default babel config for babel-loader query', t => {
+test('it returns a default babel config for babel-loader query', (t) => {
   const program = programStub('site-without-babelrc')
   const config = babelConfig(program)
 
@@ -17,7 +17,7 @@ test('it returns a default babel config for babel-loader query', t => {
   t.truthy(config.plugins.length)
 })
 
-test('all plugins are absolute paths to avoid babel lookups', t => {
+test('all plugins are absolute paths to avoid babel lookups', (t) => {
   const program = programStub('site-without-babelrc')
   const config = babelConfig(program)
 
@@ -25,27 +25,27 @@ test('all plugins are absolute paths to avoid babel lookups', t => {
   config.plugins.forEach(plugin => t.true(path.resolve(plugin) === plugin))
 })
 
-test('fixture can resolve plugins in gatsby directory (crawling up)', t => {
+test('fixture can resolve plugins in gatsby directory (crawling up)', (t) => {
   const program = programStub('site-with-valid-babelrc')
 
   const config = babelConfig(program)
   t.truthy(config.presets.length)
 })
 
-test('throws error when babelrc is not parseable', t => {
+test('throws error when babelrc is not parseable', (t) => {
   const program = programStub('site-with-invalid-babelrc')
 
   t.throws(() => babelConfig(program))
 })
 
-test('can read babel from packagejson', t => {
+test('can read babel from packagejson', (t) => {
   const program = programStub('site-with-valid-babelpackage')
 
   const config = babelConfig(program)
   t.truthy(config.presets.length)
 })
 
-test('when in development has hmre', t => {
+test('when in development has hmre', (t) => {
   const program = programStub('site-without-babelrc')
   const config = babelConfig(program, 'develop')
 
@@ -54,7 +54,16 @@ test('when in development has hmre', t => {
   t.true(includes(presetNames, 'babel-preset-react-hmre'))
 })
 
-test('throws when a plugin is not available', t => {
+test('throws when a plugin is not available', (t) => {
   const program = programStub('site-with-unresolvable-babelrc')
   t.throws(() => babelConfig(program))
+})
+
+test('throws when a configured preset is not availiable', (t) => {
+  const program = programStub('site-with-unresovleable-configured-preset-babelrc')
+  try {
+    babelConfig(program)
+  } catch (error) {
+    t.regex(error.message, /"npm install --save babel-preset-env"/)
+  }
 })
